moving save metrics from io to metrics utils
