+++ gtsfm/utils/metrics.py
@@ -231,3 +231,14 @@ def log_sfm_summary() -> None:
         "Averaging median_trans_dist_err: %.2f", averaging_metrics["translation_averaging_distance"]["median_error"]
     )
     logger.info("Averaging max_trans_dist_err: %.2f", averaging_metrics["translation_averaging_distance"]["max_error"])
+
+
+def save_metrics_as_json(metrics_groups: Delayed, output_dir: str) -> None:
+    """Saves the input metrics groups as JSON files using the name of the group.
+
+    Args:
+        metrics_groups: List of GtsfmMetricsGroup to be saved.
+        output_dir: Directory to save metrics to.
+    """
+    for metrics_group in metrics_groups:
+        metrics_group.save_to_json(os.path.join(output_dir, metrics_group.name + ".json"))
